Restructure nhswps module
